+++ RPG-Game/Player.h
@@ -24,7 +24,7 @@ public:
 
 	void Initialize();
 	void Load();
-	void Update(float deltaTime, Skeleton& skeleton, sf::Vector2f& mousePosition);
+	void Update(double deltaTime, Skeleton& skeleton, sf::Vector2f& mousePosition);
 	void Draw(sf::RenderWindow& window);
 };
