+++ RPG-Game/Skeleton.cpp
@@ -45,7 +45,7 @@ void Skeleton::Load() {
     }
 }
 
-void Skeleton::Update(float deltaTime) {
+void Skeleton::Update(double deltaTime) {
     if (health > 0) {
         boundingRectangle.setPosition(sprite.getPosition());
         healthText.setPosition(sprite.getPosition());
